+++ src/client.ts
@@ -1,6 +1,6 @@
 import { EventEmitter } from 'eventemitter3'
 import { Socket } from 'net'
-import { DeviceDrawProps, DeviceRegisterProps } from './device-types/api'
+import { ClientCapabilities, CompanionClient, DeviceDrawProps, DeviceRegisterProps } from './device-types/api'
 import { DEFAULT_PORT } from './lib'
 import * as semver from 'semver'
 
@@ -87,18 +87,18 @@ export type CompanionSatelliteClientEvents = {
 	deviceErrored: [{ deviceId: string; message: string }]
 }
 
-export class CompanionSatelliteClient extends EventEmitter<CompanionSatelliteClientEvents> {
+export class CompanionSatelliteClient extends EventEmitter<CompanionSatelliteClientEvents> implements CompanionClient {
 	private readonly debug: boolean
 	private socket: Socket | undefined
 
 	private receiveBuffer = ''
 
-	private _pingInterval: NodeJS.Timer | undefined
+	private _pingInterval: NodeJS.Timeout | undefined
 	private _pingUnackedCount = 0
 	private _lastReceivedAt = 0
 	private _connected = false
 	private _connectionActive = false // True when connected/connecting/reconnecting
-	private _retryConnectTimeout: NodeJS.Timer | undefined = undefined
+	private _retryConnectTimeout: NodeJS.Timeout | undefined = undefined
 	private _host = ''
 	private _port = DEFAULT_PORT
 	private _supportsCombinedEncoders = false
@@ -115,20 +115,15 @@ export class CompanionSatelliteClient extends EventEmitter<CompanionSatelliteCli
 	public get port(): number {
 		return this._port
 	}
-
-	/**
-	 * Until 2.4 of Companion it does not support rotary encoders.
-	 * For these, we can 'simulate' them by use the press/release actions of a button.
-	 */
-	public get useCombinedEncoders(): boolean {
-		return !this.forceSplitEncoders && this._supportsCombinedEncoders
+	public get connected(): boolean {
+		return this._connected
 	}
 
-	/**
-	 * Until 3.x of Companion it only supports providing 72x72px bitmaps for buttons.
-	 */
-	public get useCustomBitmapResolution(): boolean {
-		return this._supportsBitmapResolution
+	public get capabilities(): ClientCapabilities {
+		return {
+			useCombinedEncoders: !this.forceSplitEncoders && this._supportsCombinedEncoders,
+			useCustomBitmapResolution: this._supportsBitmapResolution,
+		}
 	}
 
 	constructor(options: CompanionSatelliteClientOptions = {}) {
@@ -196,7 +191,7 @@ export class CompanionSatelliteClient extends EventEmitter<CompanionSatelliteCli
 				return
 			}
 
-			this.emit('connected')
+			// 'connected' gets emitted once we receive 'Begin'
 		})
 
 		if (this._host) {
@@ -335,6 +330,11 @@ export class CompanionSatelliteClient extends EventEmitter<CompanionSatelliteCli
 				console.log('Companion supports bitmap resolution')
 			}
 		}
+
+		// report the connection as ready
+		setImmediate(() => {
+			this.emit('connected')
+		})
 	}
 
 	private handleState(params: Record<string, string | boolean>): void {