+++ src/device-types/api.ts
@@ -1,5 +1,3 @@
-import { CompanionSatelliteClient } from '../client'
-
 export type DeviceId = string
 
 export interface DeviceDrawProps {
@@ -25,7 +23,9 @@ export interface WrappedDevice {
 
 	close(): Promise<void>
 
-	initDevice(client: CompanionSatelliteClient, status: string): Promise<void>
+	initDevice(client: CompanionClient, status: string): Promise<void>
+
+	updateCapabilities(capabilities: ClientCapabilities): void
 
 	deviceAdded(): Promise<void>
 
@@ -36,4 +36,27 @@ export interface WrappedDevice {
 	draw(data: DeviceDrawProps): Promise<void>
 
 	showStatus(hostname: string, status: string): void
+}
+
+export interface ClientCapabilities {
+	/**
+	 * Until 2.4 of Companion it does not support rotary encoders.
+	 * For these, we can 'simulate' them by use the press/release actions of a button.
+	 */
+	readonly useCombinedEncoders: boolean
+
+	/**
+	 * Until 3.x of Companion it only supports providing 72x72px bitmaps for buttons.
+	 */
+	readonly useCustomBitmapResolution: boolean
+}
+
+export interface CompanionClient {
+	get host(): string
+
+	keyDown(deviceId: string, keyIndex: number): void
+	keyUp(deviceId: string, keyIndex: number): void
+
+	rotateLeft(deviceId: string, keyIndex: number): void
+	rotateRight(deviceId: string, keyIndex: number): void
 }